Fold all Eval functionality from the expressions into the evaluator.
