+++ Expressions/ComponentAccess.cs
@@ -41,80 +41,6 @@ namespace MetaphysicsIndustries.Solus.Expressions
         public readonly Expression Expr;
         public readonly ReadOnlyCollection<Expression> Indexes;
 
-        private int[] _evaledIndexesCache;
-        public override IMathObject Eval(SolusEnvironment env)
-        {
-            var value = Expr.Eval(env);
-            // TODO: there are some situations where we could work with a
-            //       result of Expr.Eval that is not a concrete value. for
-            //       example, "[a,2][1]" should evaluate to "2", even though
-            //       "[a,2]" with an unbound variable would not evaluate to a
-            //       concrete value.
-            switch (value)
-            {
-                case IVector v:
-                case StringValue s:
-                    if (Indexes.Count != 1)
-                        throw new OperandException(
-                            "Wrong number of indexes for the expression");
-                    break;
-                case IMatrix m:
-                    if (Indexes.Count != 2)
-                        throw new OperandException(
-                            "Wrong number of indexes for the expression");
-                    break;
-                default:
-                    throw new OperandException(
-                        "Unable to get components from expression, " +
-                        "or the expression does not have components");
-            }
-
-            if (_evaledIndexesCache == null ||
-                _evaledIndexesCache.Length < Indexes.Count)
-                _evaledIndexesCache = new int[Indexes.Count];
-            int i;
-            for (i = 0; i < Indexes.Count; i++)
-            {
-                var si = Indexes[i].Eval(env);
-                if (!(si is Number))
-                    throw new IndexException(
-                        "Indexes must be scalar");
-                var vi = si.ToNumber().Value;
-                if (!vi.IsInteger())
-                    throw new IndexException(
-                        "Indexes must be integers");
-                if (vi < 0)
-                    throw new IndexException(
-                        "Indexes must not be negative");
-                _evaledIndexesCache[i] = (int)vi;
-            }
-
-            switch (value)
-            {
-                case IVector v:
-                    if (_evaledIndexesCache[0] >= v.Length)
-                        throw new IndexException(
-                            "Index exceeds the size of the vector");
-                    return v.GetComponent(_evaledIndexesCache[0]);
-                case StringValue s:
-                    if (_evaledIndexesCache[0] >= s.Length)
-                        throw new IndexException(
-                            "Index exceeds the size of the string");
-                    return s.Value[_evaledIndexesCache[0]].ToStringValue();
-                case IMatrix m:
-                    if (_evaledIndexesCache[0] >= m.RowCount)
-                        throw new IndexException(
-                            "Index exceeds number of rows of the matrix");
-                    if (_evaledIndexesCache[1] >= m.ColumnCount)
-                        throw new IndexException(
-                            "Index exceeds number of columns of the matrix");
-                    return m.GetComponent(_evaledIndexesCache[0],
-                        _evaledIndexesCache[1]);
-            }
-
-            throw new OperandException("Unknown");
-        }
-
         public override Expression Simplify(SolusEnvironment env)
         {
             var expr = Expr.Simplify(env);
