+++ Expressions/MatrixExpression.cs
@@ -631,15 +631,6 @@ namespace MetaphysicsIndustries.Solus.Expressions
 
         #endregion
 
-        public override IMathObject Eval(SolusEnvironment env)
-        {
-            var values = new IMathObject[RowCount, ColumnCount];
-            for (int r = 0; r < RowCount; r++)
-                for (int c = 0; c < ColumnCount; c++)
-                    values[r, c] = this[r, c].Eval(env);
-            return new Matrix(values);
-        }
-
         public override Expression Clone()
         {
             var ret = new MatrixExpression(RowCount, ColumnCount);
