+++ MetaphysicsIndustries.Solus.Test/MetaphysicsIndustries.Solus.Test.csproj
@@ -94,22 +94,17 @@
     <Compile Include="EvaluatorT\FunctionsT\UnitStepFunctionT\EvalUnitStepFunctionTest.cs" />
     <Compile Include="EvaluatorT\FunctionsT\UserDefinedFunctionT\EvalUserDefinedFunctionTest.cs" />
     <Compile Include="ExpressionsT\ColorExpressionT\ResultTest.cs" />
-    <Compile Include="ExpressionsT\ComponentAccessT\EvalTest.cs" />
     <Compile Include="ExpressionsT\ComponentAccessT\ComponentAccessTest.cs" />
     <Compile Include="ExpressionsT\ComponentAccessT\SimplifyTest.cs" />
     <Compile Include="ExpressionsT\ComponentAccessT\ResultTest.cs" />
     <Compile Include="ExpressionsT\DerivativeOfVariableT\ResultTest.cs" />
     <Compile Include="ExpressionsT\FunctionCallT\ResultTest.cs" />
     <Compile Include="ExpressionsT\IntervalExpressionT\CloneTest.cs" />
-    <Compile Include="ExpressionsT\IntervalExpressionT\EvalTest.cs" />
     <Compile Include="ExpressionsT\IntervalExpressionT\ResultTest.cs" />
     <Compile Include="ExpressionsT\LiteralT\ResultTest.cs" />
-    <Compile Include="ExpressionsT\MatrixExpressionT\EvalTest.cs" />
     <Compile Include="ExpressionsT\MatrixExpressionT\ResultTest.cs" />
     <Compile Include="ExpressionsT\RandomExpressionT\ResultTest.cs" />
-    <Compile Include="ExpressionsT\VariableAccessT\EvalTest.cs" />
     <Compile Include="ExpressionsT\VariableAccessT\ResultTest.cs" />
-    <Compile Include="ExpressionsT\VectorExpressionT\EvalTest.cs" />
     <Compile Include="ExpressionsT\VectorExpressionT\ResultTest.cs" />
     <Compile Include="FunctionsT\AbsoluteValueFunctionT\AbsoluteValueFunctionTest.cs" />
     <Compile Include="FunctionsT\AbsoluteValueFunctionT\GetResultTest.cs" />
