+++ MetaphysicsIndustries.Solus.csproj
@@ -58,6 +58,7 @@
     <Compile Include="Compiler\ILCompiler.cs" />
     <Compile Include="Compiler\Instruction.cs" />
     <Compile Include="Compiler\VariableToArgumentNumberMapper.cs" />
+    <Compile Include="Evaluator.Expressions.cs" />
     <Compile Include="Exceptions\IndexException.cs" />
     <Compile Include="Exceptions\InterrogationException.cs" />
     <Compile Include="Exceptions\NameException.cs" />
